requirements.txt
# ===============================
# Core Scientific & ML Libraries
# ===============================
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
scikit-learn==1.3.0

# ===============================
# Deep Learning
# ===============================
torch==2.0.1
tensorflow==2.13.0
librosa==0.10.1
audioread==3.0.0

# ===============================
# MLOps & Experiment Tracking
# ===============================
mlflow==2.7.1
dvc==3.30.1
evidently==0.4.2

# ===============================
# API Development (Pydantic v1 stack)
# ===============================
fastapi==0.100.0
uvicorn[standard]==0.23.0
pydantic==1.10.13
# pydantic-settings is removed (conflict with Evidently & FastAPI)

# ===============================
# Data Acquisition / Networking
# ===============================
requests==2.31.0
aiohttp==3.8.5
httpx==0.24.1

# ===============================
# Data Processing & Utilities
# ===============================
python-dotenv==1.0.0
pytz==2023.3
python-dateutil==2.8.2
tqdm==4.66.1
joblib==1.3.2
PyYAML==6.0.1
psutil==5.9.5

# ===============================
# Audio Processing
# ===============================
soundfile==0.12.1
pydub==0.25.1
resampy==0.4.2

# ===============================
# Visualization
# ===============================
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
streamlit==1.28.0
folium==0.14.0
streamlit-folium==0.15.0
altair==5.1.0

# ===============================
# Monitoring & Logging
# ===============================
prometheus-client==0.17.1
loguru==0.7.2

# ===============================
# Testing
# ===============================
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# ===============================
# Database (Optional)
# ===============================
sqlalchemy==2.0.19
psycopg2-binary==2.9.7

# ===============================
# Type Checking
# ===============================
mypy==1.4.1
